QuerySortExtensions: Make more trim safe. add more xml comments
